+++ src/helpers.ts
@@ -251,7 +251,9 @@ interface FindTypes {
         | Nuke
         | Tombstone
         | Deposit
-        | Ruin;
+        | Ruin
+        | ScoreContainer
+        | ScoreCollector;
     1: RoomPosition; // FIND_EXIT_TOP
     3: RoomPosition; // FIND_EXIT_RIGHT
     5: RoomPosition; // FIND_EXIT_BOTTOM
@@ -280,6 +282,8 @@ interface FindTypes {
     121: PowerCreep; // FIND_HOSTILE_POWER_CREEPS
     122: Deposit; // FIND_DEPOSITS
     123: Ruin; // FIND_RUINS
+    10011: ScoreContainer; // FIND_SCORE_CONTAINERS
+    10012: ScoreCollector; // FIND_SCORE_COLLECTORS
 }
 
 interface FindPathOpts {