+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Adds Season world objects `ScoreContainer` and `ScoreCollector` along with constants ([#188](https://github.com/screepers/typed-screeps/pull/188))
+
+### Added
+
 - Adds `import` and `export` to `RoomVisual` and `MapVisual` ([#181](https://github.com/screepers/typed-screeps/pull/181))
 - Add `InterShardResourceConstant`, fix resources missing from `MarketResourceConstant` ([#182](https://github.com/screepers/typed-screeps/pull/182))